+++ jk_src/jk_lib/quicksort/quicksort.h
@@ -1,9 +1,11 @@
 #ifndef JK_QUICKSORT_H
 #define JK_QUICKSORT_H
 
+#include <stddef.h>
+
 void jk_quicksort(void *array,
-        int element_count,
-        int element_size,
+        size_t element_count,
+        size_t element_size,
         void *tmp,
         int (*compare)(void *a, void *b));
